requirements.txt
discord.py[voice]
yt-dlp
pynacl